+++ federation/entities/activitypub/models.py
@@ -1,12 +1,16 @@
 import copy
 import json
 import logging
+import re
+import traceback
 import uuid
-from datetime import timedelta
+from operator import attrgetter
 from typing import List, Dict, Union
-from urllib.parse import urlparse
+from unicodedata import normalize
+from urllib.parse import unquote, urlparse
 
 import bleach
+from bs4 import BeautifulSoup
 from calamus import fields
 from calamus.schema import JsonLDAnnotation, JsonLDSchema, JsonLDSchemaOpts
 from calamus.utils import normalize_value
@@ -31,10 +35,10 @@ from federation.utils.text import with_slash, validate_handle
 logger = logging.getLogger("federation")
 
 
-def get_profile_or_entity(fid):
-    obj = get_profile(fid=fid)
-    if not obj:
-        obj = retrieve_and_parse_document(fid)
+def get_profile_or_entity(**kwargs):
+    obj = get_profile(**kwargs)
+    if not obj and kwargs.get('fid'):
+        obj = retrieve_and_parse_document(kwargs['fid'])
     return obj
     
 
@@ -57,6 +61,7 @@ as2 = fields.Namespace("https://www.w3.org/ns/activitystreams#")
 dc = fields.Namespace("http://purl.org/dc/terms/")
 diaspora = fields.Namespace("https://diasporafoundation.org/ns/")
 ldp = fields.Namespace("http://www.w3.org/ns/ldp#")
+lemmy = fields.Namespace("https://join-lemmy.org/ns#")
 litepub = fields.Namespace("http://litepub.social/ns#")
 misskey = fields.Namespace("https://misskey-hub.net/ns#")
 ostatus = fields.Namespace("http://ostatus.org#")
@@ -241,8 +246,8 @@ class Object(BaseEntity, metaclass=JsonLDAnnotation):
                         metadata={'ctx':[{ 'alsoKnownAs':{'@id':'as:alsoKnownAs','@type':'@id'}}]})
     icon = MixedField(as2.icon, nested='ImageSchema')
     image = MixedField(as2.image, nested='ImageSchema')
-    tag_objects = MixedField(as2.tag, nested=['HashtagSchema','MentionSchema','PropertyValueSchema','EmojiSchema'], many=True)
-    attachment = fields.Nested(as2.attachment, nested=['ImageSchema', 'AudioSchema', 'DocumentSchema','PropertyValueSchema','IdentityProofSchema'],
+    tag_objects = MixedField(as2.tag, nested=['NoteSchema', 'HashtagSchema','MentionSchema','PropertyValueSchema','EmojiSchema'], many=True)
+    attachment = fields.Nested(as2.attachment, nested=['LinkSchema', 'NoteSchema', 'ImageSchema', 'AudioSchema', 'DocumentSchema','PropertyValueSchema','IdentityProofSchema'],
                                many=True, default=[])
     content_map = LanguageMap(as2.content)  # language maps are not implemented in calamus
     context = fields.RawJsonLD(as2.context)
@@ -250,7 +255,7 @@ class Object(BaseEntity, metaclass=JsonLDAnnotation):
     generator = MixedField(as2.generator, nested=['ApplicationSchema','ServiceSchema'])
     created_at = fields.DateTime(as2.published, add_value_types=True)
     replies = MixedField(as2.replies, nested=['CollectionSchema','OrderedCollectionSchema'])
-    signature = MixedField(sec.signature, nested = 'SignatureSchema',
+    signature = MixedField(sec.signature, nested = 'RsaSignature2017Schema',
                            metadata={'ctx': [CONTEXT_SECURITY,
                                              {'RsaSignature2017':'sec:RsaSignature2017'}]})
     start_time = fields.DateTime(as2.startTime, add_value_types=True)
@@ -332,6 +337,20 @@ class Object(BaseEntity, metaclass=JsonLDAnnotation):
             ctx = copy.copy(data['@context'])
             data['@context'] = context_manager.merge_context(ctx)
             return data
+
+        # JSONLD specs states it is case sensitive.
+        # Ensure type names for which we have an implementation have the proper case
+        # for platforms that ignore the spec.
+        @pre_load
+        def patch_types(self, data, **kwargs):
+            def walk_payload(payload):
+                for key,val in copy.copy(payload).items():
+                    if isinstance(val, dict):
+                        walk_payload(val)
+                    if key == 'type':
+                        payload[key] = MODEL_NAMES.get(val.lower(), val)
+                return payload
+            return walk_payload(data)
 
         # A node without an id isn't true json-ld, but many payloads have
         # id-less nodes. Since calamus forces random ids on such nodes, 
@@ -567,7 +586,8 @@ class Person(Object, base.Profile):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._allowed_children += (PropertyValue, IdentityProof)
+        self._required += ['url']
+        self._allowed_children += (Note, PropertyValue, IdentityProof)
 
     # Set finger to username@host if not provided by the platform
     def post_receive(self):
@@ -576,12 +596,15 @@ class Person(Object, base.Profile):
             self.finger = profile.finger
         else:
             domain = urlparse(self.id).netloc
-            finger = f'{self.username.lower()}@{domain}'
+            finger = f'{self.username}@{domain}'
             if get_profile_id_from_webfinger(finger) == self.id:
                 self.finger = finger
         # multi-protocol platform
         if self.finger and self.guid is not missing and self.handle is missing:
             self.handle = self.finger
+        # Some platforms don't set this property.
+        if self.url is missing:
+            self.url = self.id
 
     def to_as2(self):
         self.followers = f'{with_slash(self.id)}followers/'
@@ -716,15 +739,19 @@ class Note(Object, RawContentMixin):
 
     _cached_raw_content = ''
     _cached_children = []
+    _soup = None
     signable = True
 
     def __init__(self, *args, **kwargs):
         self.tag_objects = [] # mutable objects...
         super().__init__(*args, **kwargs)
-        self._allowed_children += (base.Audio, base.Video)
+        self.raw_content  # must be "primed" with source property for inbound payloads
+        self.rendered_content # must be "primed" with content_map property for inbound payloads
+        self._allowed_children += (base.Audio, base.Video, Link)
+        self._required.remove('raw_content')
+        self._required += ['rendered_content']
 
     def to_as2(self):
-        self.sensitive = 'nsfw' in self.tags
         self.url = self.id
 
         edited = False
@@ -752,8 +779,8 @@ class Note(Object, RawContentMixin):
 
     def to_base(self):
         kwargs = get_base_attributes(self, keep=(
-            '_mentions', '_media_type', '_rendered_content', '_source_object',
-            '_cached_children', '_cached_raw_content'))
+            '_mentions', '_media_type', '_source_object',
+            '_cached_children', '_cached_raw_content', '_soup'))
         entity = Comment(**kwargs) if getattr(self, 'target_id') else Post(**kwargs)
         # Plume (and maybe other platforms) send the attrbutedTo field as an array
         if isinstance(entity.actor_id, list): entity.actor_id = entity.actor_id[0]
@@ -764,6 +791,7 @@ class Note(Object, RawContentMixin):
     def pre_send(self) -> None:
         """
         Attach any embedded images from raw_content.
+        Add Hashtag and Mention objects (the client app must define the class tag/mention property)
         """
         super().pre_send()
         self._children = [
@@ -774,135 +802,138 @@ class Note(Object, RawContentMixin):
                 ) for image in self.embedded_images
                 ]
 
-        # Add other AP objects
-        self.extract_mentions()
-        self.content_map = {'orig': self.rendered_content}
-        self.add_mention_objects()
-        self.add_tag_objects()
+        # Add Hashtag objects
+        for el in self._soup('a', attrs={'class':'hashtag'}):
+            self.tag_objects.append(Hashtag(
+                href = el.attrs['href'],
+                name = el.text
+            ))
+            self.tag_objects = sorted(self.tag_objects, key=attrgetter('name'))
+            if el.text == '#nsfw': self.sensitive = True
+
+        # Add Mention objects
+        mentions = []
+        for el in self._soup('a', attrs={'class':'mention'}):
+            mentions.append(el.text.lstrip('@'))
+
+        mentions.sort()
+        for mention in mentions:
+            if validate_handle(mention):
+                profile = get_profile(finger__iexact=mention)
+                # only add AP profiles mentions
+                if getattr(profile, 'id', None):
+                    self.tag_objects.append(Mention(href=profile.id, name='@'+mention))
+                    # some platforms only render diaspora style markdown if it is available
+                    self.source['content'] = self.source['content'].replace(mention, '{' + mention + '}')
+
 
     def post_receive(self) -> None:
         """
-        Make linkified tags normal tags.
+        Mark linkified tags and mentions with a data-{mention, tag} attribute.
         """
         super().post_receive()
 
-        if not self.raw_content or self._media_type == "text/markdown":
+        if self._media_type == "text/markdown":
             # Skip when markdown
             return
 
-        hrefs = []
-        for tag in self.tag_objects:
-            if isinstance(tag, Hashtag):
-                if tag.href is not missing:
-                    hrefs.append(tag.href.lower())
-                elif tag.id is not missing:
-                    hrefs.append(tag.id.lower())
-        # noinspection PyUnusedLocal
-        def remove_tag_links(attrs, new=False):
-            # Hashtag object hrefs
-            href = (None, "href")
-            url = attrs.get(href, "").lower()
-            if url in hrefs:
-                return
-            # one more time without the query (for pixelfed)
-            parsed = urlparse(url)
-            url = f'{parsed.scheme}://{parsed.netloc}{parsed.path}'
-            if url in hrefs:
-                return
-
-            # Mastodon
-            rel = (None, "rel")
-            if attrs.get(rel) == "tag":
-                return
-            
-            # Friendica
-            if attrs.get(href, "").endswith(f'tag={attrs.get("_text")}'):
-                return
-
-            return attrs
-
-        self.raw_content = bleach.linkify(
-            self.raw_content,
-            callbacks=[remove_tag_links],
-            parse_email=False,
-            skip_tags=["code", "pre"],
-        )
+        self._find_and_mark_hashtags()
+        self._find_and_mark_mentions()
 
         if getattr(self, 'target_id'): self.entity_type = 'Comment'
 
-    def add_tag_objects(self) -> None:
-        """
-        Populate tags to the object.tag list.
-        """
-        try:
-            from federation.utils.django import get_configuration
-            config = get_configuration()
-        except ImportError:
-            tags_path = None
-        else:
-            if config["tags_path"]:
-                tags_path = f"{config['base_url']}{config['tags_path']}"
-            else:
-                tags_path = None
-        for tag in self.tags:
-            _tag = Hashtag(name=f'#{tag}')
-            if tags_path:
-                _tag.href = tags_path.replace(":tag:", tag)
-            self.tag_objects.append(_tag)
+    def _find_and_mark_hashtags(self):
+        hrefs = set()
+        for tag in self.tag_objects:
+            if isinstance(tag, Hashtag):
+                if tag.href is not missing:
+                    hrefs.add(tag.href.lower())
+                # Some platforms use id instead of href...
+                elif tag.id is not missing:
+                    hrefs.add(tag.id.lower())
 
-    def add_mention_objects(self) -> None:
-        """
-        Populate mentions to the object.tag list.
-        """
-        if len(self._mentions):
-            mentions = list(self._mentions)
-            mentions.sort()
-            for mention in mentions:
-                if validate_handle(mention):
-                    profile = get_profile(finger=mention)
-                    # only add AP profiles mentions
-                    if getattr(profile, 'id', None):
-                        self.tag_objects.append(Mention(href=profile.id, name='@'+mention))
-                        # some platforms only render diaspora style markdown if it is available
-                        self.source['content'] = self.source['content'].replace(mention, '{'+mention+'}')
+        for link in self._soup.find_all('a', href=True):
+            parsed = urlparse(unquote(link['href']).lower())
+            # remove the query part and trailing garbage, if any
+            path = parsed.path
+            trunc = re.match(r'(/[\w/\-]+)', parsed.path)
+            if trunc:
+                path = trunc.group()
+            url = f'{parsed.scheme}://{parsed.netloc}{path}'
+            # convert accented characters to their ascii equivalent
+            normalized_path = normalize('NFD', path).encode('ascii', 'ignore')
+            normalized_url = f'{parsed.scheme}://{parsed.netloc}{normalized_path.decode()}'
+            links = {link['href'].lower(), unquote(link['href']).lower(), url, normalized_url}
+            if links.intersection(hrefs):
+                tag = re.match(r'^#?([\w\-]+$)', link.text)
+                if tag:
+                    link['data-hashtag'] = tag.group(1).lower()
+
+    def _find_and_mark_mentions(self):
+        mentions = [mention for mention in self.tag_objects if isinstance(mention, Mention)]
+        # There seems to be consensus on using the profile url for
+        # the link and the profile id for the Mention object href property,
+        # but some platforms will set mention.href to the profile url, so
+        # we check both.
+        for mention in mentions:
+            hrefs = []
+            profile = get_profile_or_entity(fid=mention.href, remote_url=mention.href)
+            if profile and not profile.url:
+                # This should be removed when we are confident that the remote_url property
+                # has been populated for most profiles on the client app side.
+                profile = retrieve_and_parse_profile(profile.id)
+            if profile:
+                hrefs.extend([profile.id, profile.url])
+            for href in hrefs:
+                links = self._soup.find_all(href=href)
+                for link in links:
+                    link['data-mention'] = profile.finger
+                    self._mentions.add(profile.finger)
 
     def extract_mentions(self):
         """
-        Extract mentions from the source object.
+        Attempt to extract mentions from raw_content if available
         """
-        super().extract_mentions()
 
-        if getattr(self, 'tag_objects', None):
-            #tag_objects = self.tag_objects if isinstance(self.tag_objects, list) else [self.tag_objects]
-            for tag in self.tag_objects:
-                if isinstance(tag, Mention):
-                    profile = get_profile_or_entity(fid=tag.href)
-                    handle = getattr(profile, 'finger', None)
-                    if handle: self._mentions.add(handle)
+        if self.raw_content:
+            super().extract_mentions()
+        return
 
     @property
-    def raw_content(self):
-
-        if self._cached_raw_content: return self._cached_raw_content
+    def rendered_content(self):
+        if self._soup: return str(self._soup)
+        content = ''
         if self.content_map:
             orig = self.content_map.pop('orig')
             if len(self.content_map.keys()) > 1:
                 logger.warning('Language selection not implemented, falling back to default')
-                self._rendered_content = orig.strip()
+                content = orig.strip()
             else:
-                self._rendered_content = orig.strip() if len(self.content_map.keys()) == 0 else next(iter(self.content_map.values())).strip()
+                content = orig.strip() if len(self.content_map.keys()) == 0 else next(iter(self.content_map.values())).strip()
             self.content_map['orig'] = orig
+        # to allow for posts/replies with medias only.
+        if not content: content = "<div></div>"
+        self._soup = BeautifulSoup(content, 'html.parser')
+        return str(self._soup)
 
-            if isinstance(self.source, dict) and self.source.get('mediaType') == 'text/markdown':
-                self._media_type = self.source['mediaType']
-                self._cached_raw_content = self.source.get('content').strip()
-            else:
-                self._media_type = 'text/html'
-                self._cached_raw_content = self._rendered_content
-            # to allow for posts/replies with medias only.
-            if not self._cached_raw_content: self._cached_raw_content = "<div></div>"
-            return self._cached_raw_content
-    
+    @rendered_content.setter
+    def rendered_content(self, value):
+        if not value: return
+        self._soup = BeautifulSoup(value, 'html.parser')
+        self.content_map = {'orig': value}
+
+    @property
+    def raw_content(self):
+        if self._cached_raw_content: return self._cached_raw_content
+
+        if isinstance(self.source, dict) and self.source.get('mediaType') == 'text/markdown':
+            self._media_type = self.source['mediaType']
+            self._cached_raw_content = self.source.get('content').strip()
+        else:
+            self._media_type = 'text/html'
+            self._cached_raw_content = ""
+        return self._cached_raw_content
+
     @raw_content.setter
     def raw_content(self, value):
         if not value: return
@@ -917,12 +948,13 @@ class Note(Object, RawContentMixin):
         if isinstance(getattr(self, 'attachment', None), list):
             children = []
             for child in self.attachment:
-                if isinstance(child, Document):
-                    obj = child.to_base()
-                    if isinstance(obj, Image):
-                        if obj.inline or (obj.image and obj.image in self.raw_content):
+                if isinstance(child, (Document, Link)):
+                    if hasattr(child, 'to_base'):
+                        child = child.to_base()
+                    if isinstance(child, Image):
+                        if child.inline or (child.image and child.image in self.raw_content):
                             continue
-                    children.append(obj)
+                    children.append(child)
             self._cached_children = children
 
         return self._cached_children
@@ -1010,7 +1042,7 @@ class Video(Document, base.Video):
                 self.actor_id = new_act[0]
             
             entity = Post(**get_base_attributes(self,
-                keep=('_mentions', '_media_type', '_rendered_content',
+                keep=('_mentions', '_media_type', '_soup',
                       '_cached_children', '_cached_raw_content', '_source_object')))
             set_public(entity)
             return entity
@@ -1019,7 +1051,7 @@ class Video(Document, base.Video):
             return self
 
 
-class Signature(Object):
+class RsaSignature2017(Object):
     created = fields.DateTime(dc.created, add_value_types=True)
     creator = IRI(dc.creator)
     key = fields.String(sec.signatureValue)
@@ -1174,6 +1206,7 @@ class Retraction(Announce, base.Retraction):
 
 class Tombstone(Object, base.Retraction):
     target_id = fields.Id()
+    signable = True
 
     def to_as2(self):
         if not isinstance(self.activity, type): return None
@@ -1294,7 +1327,7 @@ def extract_receivers(entity):
     profile = None
     # don't care about receivers for payloads without an actor_id    
     if getattr(entity, 'actor_id'):
-        profile = get_profile_or_entity(entity.actor_id)
+        profile = get_profile_or_entity(fid=entity.actor_id)
     if not isinstance(profile, base.Profile):
         return receivers
     
@@ -1314,14 +1347,16 @@ def extract_and_validate(entity):
     entity._source_protocol = "activitypub"
     # Extract receivers
     entity._receivers = extract_receivers(entity)
+
+    # Extract mentions
+    if hasattr(entity, "extract_mentions"):
+        entity.extract_mentions()
+
     if hasattr(entity, "post_receive"):
         entity.post_receive()
 
     if hasattr(entity, 'validate'): entity.validate()
 
-    # Extract mentions
-    if hasattr(entity, "extract_mentions"):
-        entity.extract_mentions()
 
 
 def extract_replies(replies):
@@ -1373,6 +1408,9 @@ def element_to_objects(element: Union[Dict, Object], sender: str = "") -> List:
             logger.error("Failed to validate entity %s: %s", entity, ex)
             return []
         except InvalidSignature as exc:
+            if isinstance(entity, base.Retraction):
+                logger.warning('Relayed retraction on %s, ignoring', entity.target_id)
+                return []
             logger.info('%s, fetching from remote', exc)
             entity = retrieve_and_parse_document(entity.id)
             if not entity:
@@ -1396,6 +1434,7 @@ def model_to_objects(payload):
             entity = model.schema().load(payload)
         except (KeyError, jsonld.JsonLdError, exceptions.ValidationError) as exc :  # Just give up for now. This must be made robust
             logger.error("Error parsing jsonld payload (%s)", exc)
+            traceback.print_exception(exc)
             return None
 
         if isinstance(getattr(entity, 'object_', None), Object):
@@ -1417,3 +1456,9 @@ CLASSES_WITH_CONTEXT_EXTENSIONS = (
     PropertyValue
 )
 context_manager = LdContextManager(CLASSES_WITH_CONTEXT_EXTENSIONS)
+
+
+MODEL_NAMES = {}
+for key,val in copy.copy(globals()).items():
+    if type(val) == JsonLDAnnotation and issubclass(val, (Object, Link)):
+        MODEL_NAMES[key.lower()] = key